tests/baseline/protocol-1.2/inspector/console/console-filter-level-test-expected.txt

CONSOLE INFO: line 10: sample info
CONSOLE MESSAGE: line 11: sample log
CONSOLE WARNING: line 12: sample warning
CONSOLE DEBUG: line 13: sample debug
CONSOLE ERROR: line 14: sample error
CONSOLE INFO: line 16: abc info
CONSOLE INFO: line 17: def info
CONSOLE WARNING: line 19: abc warn
CONSOLE WARNING: line 20: def warn
Tests that console can filter messages by source.


Running: beforeFilter
beforeFilter
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: verbose
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:8 sample debug
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: info
>console-filter-level-test.html:5 sample info
>console-filter-level-test.html:6 sample log
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:12 def info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: warning
>console-filter-level-test.html:7 sample warning
onload @ console-filter-level-test.html:7
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>console-filter-level-test.html:15 def warn
onload @ console-filter-level-test.html:15
>'Should be always visible'
>"Should be always visible"

Running: error
>console-filter-level-test.html:9 sample error
onload @ console-filter-level-test.html:9
>'Should be always visible'
>"Should be always visible"

Running: abcMessagePlain
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>'Should be always visible'
>"Should be always visible"

Running: abcMessageRegex
>console-filter-level-test.html:11 abc info
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>'Should be always visible'
>"Should be always visible"

Running: abcMessageRegexWarning
>console-filter-level-test.html:14 abc warn
onload @ console-filter-level-test.html:14
>'Should be always visible'
>"Should be always visible"
